% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_radar.R
\name{create_radar}
\alias{create_radar}
\title{Plot an interactive radar/spider-chart of the ESD data towards a given
reference}
\usage{
create_radar(ESD_data, reference, threshold = 0.5)
}
\arguments{
\item{ESD_data}{tidy version of the ESD data representation}

\item{threshold}{provide the value on which a red threshold line should be
provided}

\item{char}{choose an option from: wug, gemeente, vlaanderen,
wug_gemeente, wug_provincie, wug_vlaanderen}
}
\description{
Plot an interactive radar/spider-chart of the ESD data towards a given
reference
}

