% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{get_locations}
\alias{get_locations}
\title{get from link-table containing an overview of the WUG-NR - Provincie -
GEMEENTE relation the corresponding province and municipality based on the
WUG identifier}
\usage{
get_locations(link_table, id_wug)
}
\arguments{
\item{link_table}{data.frame}

\item{id_wug}{char identifier for the WUG}
}
\value{
data.frame containing the wug municipality and province
}
\description{
get from link-table containing an overview of the WUG-NR - Provincie -
GEMEENTE relation the corresponding province and municipality based on the
WUG identifier
}

