% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{get_wug_ids}
\alias{get_wug_ids}
\title{get from link-table containing an overview of the WUG-NR - Provincie -
GEMEENTE relation the corresponding province and municipality based on the
WUG identifier}
\usage{
get_wug_ids(link_table)
}
\arguments{
\item{link_table}{data.frame}

\item{id_wug}{char identifier for the WUG}
}
\value{
list with unique wug identifier codes
}
\description{
get from link-table containing an overview of the WUG-NR - Provincie -
GEMEENTE relation the corresponding province and municipality based on the
WUG identifier
}

